+++ src/app/app.ts
@@ -24,7 +24,7 @@ import GAMEPAD_MAP from "../data/gamepad.json";
 import {
     type ButtonBinding,
     type ButtonsDef,
-    getLastDevice,
+    getLastInputType,
 } from "./inputBindings";
 
 export class ButtonState<T = string> {
@@ -1211,7 +1211,7 @@ export const initApp = (opt: {
         onButtonPress,
         onButtonDown,
         onButtonRelease,
-        getLastDevice,
+        getLastInputType,
         events: state.events,
     };
 };